Slight Optimization and Organization
Removed some buttons that we are not going to implement. Added a button to skip the sign in proccess for now just to get to the main page to start work there. Debugging Google and email login, SHA-1 and API exception errors are the main problem here, something that can be sorted out as we go along. Implementation basically finished.