User information page is partialy implemented
Added a customer usage page, waiting upon the DB to be implemented to finish implemenation here (need the user table atleast). New screen can be found under the Settings tab/button. There is a non-functional Darkmode button there too, just need to mess with the styles to get that to work as far as I can tell. Also did minor changes in main activity to attemt to resolve a smaller problem on startup when there is not user signed into the app, still a WIP.