Started going through receipts
Started going through many receipts examples from online to add to the number of terms that we want to through out and which to look at for keywords. I will probably need to write a regex to help parse out data that is not needed to keep (number sequences that are not prices) among other things. Will need help with getting this together