Integrated DB and ML
Got the DB from Nesi, no idea if it works as I had to make changes to it on the fly to correct values. Items in the DB that should have been Doubles were Ints, methods are lacking certain input values (adding an item doesn't let you specify an amount of said item). More things like this.