Fixed and implemented some stuff
Fixed the graphical error in the wheel (now dynamically updates both categories and amounts), removed the receipt adapter and moved it into the history fragment file. Implemented the set on click listener for history fragments. Need to fix a bug in the DB code for Receipt Update and also fix the edit button.